Переписала часть кода; разделила подробнее архитектуру; Заменила логику таймера на momentJS.
