--- src/components/Carousel.js
@@ -1,17 +1,0 @@
-import React from 'react'
-import CarouselContainer, { CarouselItem } from './CarouselContainer'
-
-function Carousel() {
-  return (
-    <CarouselContainer>
-      <CarouselItem>Слайд 1</CarouselItem>
-      <CarouselItem>Слайд 2</CarouselItem>
-      <CarouselItem>Слайд 3</CarouselItem>
-      <CarouselItem>Слайд 4</CarouselItem>
-      <CarouselItem>Слайд 5</CarouselItem>
-      <CarouselItem>Слайд 6</CarouselItem>
-    </CarouselContainer>
-  )
-}
-
-export default Carousel
