+++ src/components/accordion/style.js
@@ -1,0 +1,27 @@
+import styled from 'styled-components'
+
+export const AccordionSection = styled.section`
+  background-color: #ffe7e7;
+  padding: 120px 0;
+`
+
+export const Container = styled.div`
+  max-width: 1270px;
+  width: 100%;
+  height: 100%;
+  padding: 0 15px;
+  margin-inline: auto;
+`
+
+export const Title = styled.h2`
+  font-weight: 600;
+  font-size: 36px;
+  line-height: 44px;
+  margin: 0 0 40px 0;
+`
+
+export const AccordionWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+`
