+++ src/components/footer/footer.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import Container from '../style_cont'
+import { FooterEl, FooterText, Wrapper } from './style'
+
+const Footer = () => (
+  <FooterEl>
+    <Container>
+      <Wrapper>
+        <FooterText>Lorem ipsum dolor sit amet consectetur</FooterText>
+        <FooterText>Lorem ipsum dolor sit amet consectetur</FooterText>
+        <FooterText>Lorem ipsum dolor sit amet consectetur</FooterText>
+      </Wrapper>
+    </Container>
+  </FooterEl>
+)
+
+export default Footer;
