+++ src/components/footer/style.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components'
+
+export const FooterEl = styled.footer`
+  background: #D2D2D2;
+`
+
+export const Wrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 2rem 0;
+`
+
+export const FooterText = styled.p`
+  width: 25%;
+  text-align: center;
+  background: #C0C0C0;
+`
+
