+++ src/components/instruction/style.js
@@ -1,0 +1,63 @@
+import styled from 'styled-components'
+
+export const InstrSec = styled.section`
+  background: #FFFFFF;
+`
+
+export const InstrHeader = styled.p`
+  display: inline-block;
+  font-weight: 600;
+  font-size: 36px;
+  line-height: 44px;
+  text-align: center;
+  margin-top: 96px;
+`
+
+export const Wrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`
+
+export const InstrList = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 60%;
+  margin-top: 160px;
+`
+
+export const InstrNums = styled.ul`
+  margin: 20px 0 0;
+  display: inline-block;
+  font-size: 80px;
+  list-style-type: none;
+`
+
+export const InstrNum = styled.li`
+  background: #C0C0C0;
+  margin: 10px 0 163px;
+  padding-left: 20px;
+  padding-right: 20px;
+`
+
+export const InstrText = styled.ul`
+  width: 100%;
+  text-align: center;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 35px;
+  list-style-type: none;
+`
+
+export const InstrTextLine = styled.li`
+  margin-left: 47px;
+  text-align: left;
+  margin-top: 10px;
+  background-color: #C0C0C0;
+  margin-bottom: 163px;
+  padding-left: 15px;
+`
+
+export const InstrIllustr = styled.div`
+  min-height: 586px;
+`
