+++ src/components/style_cont.js
@@ -2,11 +2,11 @@ import styled from 'styled-components'
 
 // Контейнер HTML
 
-const Cont = styled.div`
+const Container = styled.div`
   width: 100%;
   max-width: 1270px;
   margin: 0 auto;
   padding: 0 15px;
 `
 
-export default Cont
+export default Container
