--- src/components/style_global.js
@@ -1,11 +1,0 @@
-import { createGlobalStyle } from 'styled-components'
-
-const GlStyles = createGlobalStyle`
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-`
-
-export default GlStyles
