+++ src/constants/index.js
@@ -1,0 +1,35 @@
+// Временные вопросы и ответы
+const DATA_FAQ = [
+  {
+    id: 1,
+    question: 'Lorem ipsum dolor sit amet, consectetur',
+    answer:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Sem fringilla ut morbi tincidunt augue interdum. Morbi tristique senectus et netus et malesuada fames. Adipiscing tristique risus nec feugiat in fermentum '
+  },
+  {
+    id: 2,
+    question: 'Lorem ipsum dolor sit amet, consectetur',
+    answer:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Sem fringilla ut morbi tincidunt augue interdum. Morbi tristique senectus et netus et malesuada fames. Adipiscing tristique risus nec feugiat in fermentum '
+  },
+  {
+    id: 3,
+    question: 'Lorem ipsum dolor sit amet, consectetur',
+    answer:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Sem fringilla ut morbi tincidunt augue interdum. Morbi tristique senectus et netus et malesuada fames. Adipiscing tristique risus nec feugiat in fermentum '
+  },
+  {
+    id: 4,
+    question: 'Lorem ipsum dolor sit amet, consectetur',
+    answer:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Sem fringilla ut morbi tincidunt augue interdum. Morbi tristique senectus et netus et malesuada fames. Adipiscing tristique risus nec feugiat in fermentum '
+  },
+  {
+    id: 5,
+    question: 'Lorem ipsum dolor sit amet, consectetur',
+    answer:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Sem fringilla ut morbi tincidunt augue interdum. Morbi tristique senectus et netus et malesuada fames. Adipiscing tristique risus nec feugiat in fermentum '
+  }
+]
+
+export default DATA_FAQ
