+++ src/pages/auth/register.js
