+++ src/pages/box-create/box-styles.js
@@ -1,4 +1,5 @@
 import styled from 'styled-components'
+import { Switch } from "@mui/material";
 
 export const Container = styled.div`
   display: flex;
@@ -9,6 +10,7 @@ export const Container = styled.div`
   margin: 0 auto;
   padding-bottom: 80px;
 `
+
 export const Head = styled.div`
   display: flex;
   flex-direction: column;
@@ -16,6 +18,7 @@ export const Head = styled.div`
   align-items: center;
   margin-bottom: 87px;
 `
+
 export const Title = styled.h1`
   font-family: 'Inter';
   font-style: normal;
@@ -25,6 +28,7 @@ export const Title = styled.h1`
   color: #000000;
   margin-bottom: 12px;
 `
+
 export const SubTitle = styled.p`
   font-family: 'Inter';
   font-style: normal;
@@ -57,9 +61,11 @@ export const Input = styled.input`
   text-align: center;
   margin-bottom: 50px;
 `
+
 export const DivInput = styled.div`
   position: relative;
 `
+
 export const Label = styled.label`
   position: absolute;
   font-family: 'Inter';
@@ -71,6 +77,7 @@ export const Label = styled.label`
   top: -30px;
   left: 0;
 `
+
 export const P = styled.p`
   font-family: 'Inter';
   font-style: normal;
@@ -79,6 +86,7 @@ export const P = styled.p`
   line-height: 22px;
   color: #000000;
 `
+
 export const Cover = styled.div`
   margin-bottom: 85px;
 `
@@ -90,10 +98,12 @@ export const CoverButton = styled.button`
   border-radius: 3px;
   font-size: 58px;
 `
+
 export const SmallInput = styled(Input)`
   height: 50px;
   margin-bottom: 46px;
 `
+
 export const SmallLabel = styled(Label)`
   font-weight: 400;
   font-size: 13px;
@@ -114,14 +124,17 @@ export const CreateButton = styled.button`
   line-height: 29px;
   color: #000000;
 `
+
 export const CancellButton = styled(CreateButton)`
   background: #d9d9d9;
   color: #9b9b9b;
 `
+
 export const ButtonsDiv = styled.div`
   display: flex;
   justify-content: space-between;
 `
+
 export const Select = styled.select`
   width: 304px;
   height: 61px;
@@ -134,6 +147,7 @@ export const Select = styled.select`
   line-height: 24px;
   color: #000000;
 `
+
 export const CostDiv = styled(ButtonsDiv)`
   margin-bottom: 85px;
 `
@@ -150,3 +164,47 @@ export const CostInput = styled.input`
   line-height: 29px;
   color: #000000;
 `
+
+
+export const AntSwitch = styled(Switch)(({ theme }) => ({
+  width: 79,
+  height: 36,
+  padding: 0,
+  display: 'flex',
+  '&:active': {
+    '& .MuiSwitch-thumb': {
+      width: 25
+    },
+    '& .MuiSwitch-switchBase.Mui-checked': {
+      transform: 'translateX(9px)'
+    }
+  },
+  '& .MuiSwitch-switchBase': {
+    padding: 3,
+    '&.Mui-checked': {
+      transform: 'translateX(43px)',
+      color: '#FD9797',
+      '& + .MuiSwitch-track': {
+        opacity: 1,
+        backgroundColor: '#FFBFBF'
+      }
+    }
+  },
+  '& .MuiSwitch-thumb': {
+    width: 30,
+    height: 30,
+    borderRadius: 15,
+    transition: theme.transitions.create(['width'], {
+      duration: 200
+    })
+  },
+  '& .MuiSwitch-track': {
+    borderRadius: 38 / 2,
+    opacity: 1,
+    backgroundColor:
+      theme.palette.mode === 'dark'
+        ? 'rgba(255,255,255,.35)'
+        : 'rgba(0,0,0,.25)',
+    boxSizing: 'border-box'
+  }
+}))
